requirements/base.txt
alembic==1.2.1
boto3==1.12.38
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
email-validator==1.0.5
bcrypt==3.1.7
fastapi==0.52.0
h11==0.9.0
httptools==0.1.1
idna==2.9
Mako==1.1.2
MarkupSafe==1.1.1
passlib==1.7.2
python-multipart==0.0.5
psycopg2-binary==2.8.3
pydantic==1.4
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
python-json-logger==0.1.11
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.8
starlette==0.13.2
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
python-decouple==3.3
